Rename follow base presenter and split presenter tests
