+++ pages/src/components/follow/follow-live-overlay-viewer/follow-live-overlay-viewer-presenter.ts
@@ -1,5 +1,5 @@
 import type { FollowLiveOverlayPresentArgs, FollowLiveOverlayPresentation } from "../types";
-import { FollowLiveBasePresenter } from "../follow-live-presenter";
+import { FollowLiveBasePresenter } from "../follow-live-base-presenter";
 
 export class FollowLiveOverlayViewerPresenter extends FollowLiveBasePresenter {
   public present(args: FollowLiveOverlayPresentArgs): FollowLiveOverlayPresentation {
