+++ pages/src/components/follow/follow-live-overlay-viewer/tests/follow-live-overlay-viewer-presenter.test.ts
@@ -1,0 +1,41 @@
+import { describe, expect, it } from "vitest";
+import type { TrackerDirectory } from "@guilty-spark/shared/contracts/individual-tracker/follow";
+import { aDirectoryWith, aTrackerWith } from "@guilty-spark/shared/contracts/individual-tracker/fakes/follow.fake";
+import { FollowLiveOverlayViewerPresenter } from "../follow-live-overlay-viewer-presenter";
+
+describe("FollowLiveOverlayViewerPresenter", () => {
+  it("falls back to an isLive tracker when liveTrackerId does not match", () => {
+    const directory: TrackerDirectory = aDirectoryWith({
+      trackers: [
+        aTrackerWith({ trackerId: "tracker-1", gamertag: "Spartan One", isLive: false }),
+        aTrackerWith({ trackerId: "tracker-2", gamertag: "Spartan Two", isLive: true }),
+      ],
+      liveTrackerId: "tracker-missing",
+    });
+    const presenter = new FollowLiveOverlayViewerPresenter();
+
+    const overlay = presenter.present({
+      gamertag: "Streamer",
+      directory,
+    });
+
+    expect(overlay.liveTracker?.trackerId).toBe("tracker-2");
+    expect(overlay.liveTrackerView?.trackerId).toBe("tracker-2");
+    expect(overlay.title).toBe("Streamer overlay - Spartan Two live - Guilty Spark");
+  });
+
+  it("returns non-live overlay title when there is no live tracker", () => {
+    const directory: TrackerDirectory = aDirectoryWith({
+      trackers: [aTrackerWith({ trackerId: "tracker-1", gamertag: "Spartan One", isLive: false })],
+      liveTrackerId: null,
+    });
+    const presenter = new FollowLiveOverlayViewerPresenter();
+
+    const overlay = presenter.present({
+      gamertag: "Streamer",
+      directory,
+    });
+
+    expect(overlay.title).toBe("Streamer overlay - Guilty Spark");
+  });
+});
