+++ pages/src/components/follow/follow-live-viewer/tests/follow-live-viewer-presenter.test.ts
@@ -1,8 +1,7 @@
 import { describe, expect, it } from "vitest";
 import type { TrackerDirectory } from "@guilty-spark/shared/contracts/individual-tracker/follow";
 import { aDirectoryWith, aTrackerWith } from "@guilty-spark/shared/contracts/individual-tracker/fakes/follow.fake";
-import { FollowLiveOverlayViewerPresenter } from "../follow-live-overlay-viewer/follow-live-overlay-viewer-presenter";
-import { FollowLiveViewerPresenter } from "../follow-live-viewer/follow-live-viewer-presenter";
+import { FollowLiveViewerPresenter } from "../follow-live-viewer-presenter";
 
 describe("FollowLiveViewerPresenter", () => {
   it("presents selected tracker view with directory streamer settings", () => {
@@ -77,43 +76,5 @@ describe("FollowLiveViewerPresenter", () => {
     });
 
     expect(viewer.title).toBe("Streamer live view - Guilty Spark");
-  });
-});
-
-describe("FollowLiveOverlayViewerPresenter", () => {
-
-  it("falls back to an isLive tracker when liveTrackerId does not match", () => {
-    const directory: TrackerDirectory = aDirectoryWith({
-      trackers: [
-        aTrackerWith({ trackerId: "tracker-1", gamertag: "Spartan One", isLive: false }),
-        aTrackerWith({ trackerId: "tracker-2", gamertag: "Spartan Two", isLive: true }),
-      ],
-      liveTrackerId: "tracker-missing",
-    });
-    const presenter = new FollowLiveOverlayViewerPresenter();
-
-    const overlay = presenter.present({
-      gamertag: "Streamer",
-      directory,
-    });
-
-    expect(overlay.liveTracker?.trackerId).toBe("tracker-2");
-    expect(overlay.liveTrackerView?.trackerId).toBe("tracker-2");
-    expect(overlay.title).toBe("Streamer overlay - Spartan Two live - Guilty Spark");
-  });
-
-  it("returns non-live overlay title when there is no live tracker", () => {
-    const directory: TrackerDirectory = aDirectoryWith({
-      trackers: [aTrackerWith({ trackerId: "tracker-1", gamertag: "Spartan One", isLive: false })],
-      liveTrackerId: null,
-    });
-    const presenter = new FollowLiveOverlayViewerPresenter();
-
-    const overlay = presenter.present({
-      gamertag: "Streamer",
-      directory,
-    });
-
-    expect(overlay.title).toBe("Streamer overlay - Guilty Spark");
   });
 });
